Add board_edit language

+++ admin/delete_message.php
@@ -3,16 +3,24 @@ header('content-type:text/html;charset=utf-8');
 require dirname(__FILE__).'/header_command.php';
 
 if (isset($_POST['delete_message']) && $now_admin == 1) {
-    $delete = $_POST['checkbox'];
-    foreach($delete as $selected_massage) {
-        $delete_sql = 'DELETE FROM msg WHERE id = '.$selected_massage;
-        $con->query($delete_sql);
+    if(!empty($_POST['checkbox'])) {
+        $delete = $_POST['checkbox'];
+        foreach($delete as $selected_massage) {
+            $delete_sql = 'DELETE FROM msg WHERE id = '.$selected_massage;
+            $con->query($delete_sql);
+        }
+        echo '
+        <script>
+          alert("'.$lang_delete_success.'");location.href="../admin.php";
+        </script>
+        ';
+    } else {
+        echo '
+        <script>
+          alert("'.$lang_delete_empty.'");location.href="../admin.php";
+        </script>
+        ';
     }
-    echo '
-    <script>
-      alert("'.$lang_delete_success.'");location.href="../";
-    </script>
-    ';
 } else {
     echo '
     <script>
